+++ test/equatable_test.dart
@@ -1,9 +1,8 @@
 import 'dart:convert';
 
+import 'package:equatable/equatable.dart';
 import 'package:equatable/src/equatable_utils.dart';
 import 'package:test/test.dart';
-
-import 'package:equatable/equatable.dart';
 
 import 'custom_list.dart';
 
@@ -90,6 +89,30 @@ class Credentials extends Equatable {
 
   @override
   List<Object> get props => [username, password];
+}
+
+class ComplexStringify extends Equatable {
+  final String name;
+  final int age;
+  final Color hairColor;
+
+  ComplexStringify({this.name, this.age, this.hairColor});
+
+  @override
+  List get props => [name, age, hairColor];
+
+  @override
+  bool get stringify => true;
+}
+
+class NullProps extends Equatable {
+  NullProps();
+
+  @override
+  List get props => null;
+
+  @override
+  bool get stringify => true;
 }
 
 void main() {
@@ -728,6 +751,37 @@ void main() {
         expect(instanceA != instanceB, true);
         expect(instanceA.hashCode != instanceB.hashCode, true);
       });
+    });
+  });
+
+  group('To String Equatable', () {
+    test('with Complex stringify', () {
+      final instanceA = ComplexStringify();
+      final instanceB = ComplexStringify(name: "Bob", hairColor: Color.black);
+      final instanceC =
+          ComplexStringify(name: "Joe", age: 50, hairColor: Color.blonde);
+      expect(instanceA.toString(), 'ComplexStringify(, , )');
+      expect(instanceB.toString(), 'ComplexStringify(Bob, , Color.black)');
+      expect(instanceC.toString(), 'ComplexStringify(Joe, 50, Color.blonde)');
+    });
+  });
+
+  group('Null props Equatable', () {
+    test('should not crash invoking equals method', () {
+      final instanceA = NullProps();
+      final instanceB = NullProps();
+      expect(instanceA == instanceB, true);
+    });
+
+    test('should not crash invoking hascode method', () {
+      final instanceA = NullProps();
+      final instanceB = NullProps();
+      expect(instanceA.hashCode == instanceB.hashCode, true);
+    });
+
+    test('should not crash invoking toString method', () {
+      final instance = NullProps();
+      expect(instance.toString(), 'NullProps()');
     });
   });
 }